refactor: Split `BottomButtonsStripe` as its own component
